fix(Tabs): fix lint
