+++ packages/react/src/components/tabs/tab-button.test.tsx
@@ -4,7 +4,12 @@ import { mountWithProviders } from '../../test-utils/renderer';
 import { TabButton } from './tab-button';
 
 describe('TabButton', () => {
+    const defaultProps = {
+        size: 'default' as const,
+    };
+
     const focusedAndSelected = {
+        ...defaultProps,
         id: 'aId',
         panelId: 'aPanelId',
         isSelected: true,
@@ -15,7 +20,7 @@ describe('TabButton', () => {
 
     test('should display button text', () => {
         const expectedButtonText = 'some text';
-        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected}>{expectedButtonText}</TabButton>);
+        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected} size="default">{expectedButtonText}</TabButton>);
 
         const tabPanel = getByTestId(wrapper, 'tab-text');
 
@@ -23,7 +28,7 @@ describe('TabButton', () => {
     });
 
     test('should not have a left icon in button when tab doesn\'t have a left icon name', () => {
-        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected}>some text</TabButton>);
+        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected} size="default">some text</TabButton>);
 
         const tabButtonLeftIcon = findByTestId(wrapper, 'tab-left-icon');
 
@@ -33,7 +38,7 @@ describe('TabButton', () => {
     test('should have a left icon in button when tab has a left icon name', () => {
         const expectedLeftIcon = 'chevronUp';
         const wrapper = mountWithProviders(
-            <TabButton {...focusedAndSelected} leftIcon={expectedLeftIcon}>some text</TabButton>,
+            <TabButton {...focusedAndSelected} size="default" leftIcon={expectedLeftIcon}>some text</TabButton>,
         );
 
         const tabButtonLeftIcon = getByTestId(wrapper, 'tab-left-icon');
@@ -42,7 +47,7 @@ describe('TabButton', () => {
     });
 
     test('should not have a right icon in button when tab doesn\'t have a right icon name', () => {
-        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected}>some text</TabButton>);
+        const wrapper = mountWithProviders(<TabButton {...focusedAndSelected} size="default">some text</TabButton>);
 
         const tabButtonRightIcon = findByTestId(wrapper, 'tab-right-icon');
 
@@ -52,7 +57,7 @@ describe('TabButton', () => {
     test('should have a right icon in button when tab has a right icon name', () => {
         const expectedRightIcon = 'chevronDown';
         const wrapper = mountWithProviders(
-            <TabButton {...focusedAndSelected} rightIcon={expectedRightIcon}>some text</TabButton>,
+            <TabButton {...focusedAndSelected} size="default" rightIcon={expectedRightIcon}>some text</TabButton>,
         );
 
         const tabButtonRightIcon = getByTestId(wrapper, 'tab-right-icon');
@@ -68,6 +73,7 @@ describe('TabButton', () => {
                 panelId="aPanelId"
                 isSelected
                 onClick={expectedOnClickCall}
+                size="default"
             >
                 some text
             </TabButton>,
